in progress

+++ SimpleRender/Drawing/Draw3D.cs
@@ -10,7 +10,7 @@ namespace SimpleRender.Drawing
 {
     public class Draw3D
     {
-        public static void Triangle(Vector3i t0, Vector3i t1, Vector3i t2, Bitmap image, Color color, int[] zbuffer)
+        public static void Triangle(Vector3f t0, Vector3f t1, Vector3f t2, Bitmap image, Color color, int[] zbuffer)
         {
         }
     }
